Refactoring, move os and os.exec wrapper into separate packages
