+++ agent/firewall_test.go
@@ -23,6 +23,8 @@ import (
 	"net"
 	"strings"
 	"testing"
+
+	utilexec "github.com/romana/core/pkg/util/exec"
 )
 
 // TestNewChains is checking that detectMissingChains correctly detects which
@@ -30,10 +32,10 @@ import (
 func TestNewChains(t *testing.T) {
 	agent := mockAgent()
 
-	// detectMissingChains calls isChainExist which is reading FakeExecutor
+	// detectMissingChains calls isChainExist which is reading utilexec.FakeExecutor
 	// isChainExist doesn't care for output but must receive not nil error
 	// otherwise it would decide that chain exist already and skip
-	E := &FakeExecutor{nil, errors.New("bla"), nil}
+	E := &utilexec.FakeExecutor{nil, errors.New("bla"), nil}
 	agent.Helper.Executor = E
 	ip := net.ParseIP("127.0.0.1")
 	fw, _ := NewFirewall(NetIf{"eth0", "A", ip}, &agent)
@@ -45,7 +47,7 @@ func TestNewChains(t *testing.T) {
 	}
 
 	// TODO test case with some chains already exist requires support for
-	// stack of output in FakeExecutor
+	// stack of output in utilexec.FakeExecutor
 }
 
 // TestCreateChains is checking that CreateChains generates correct OS commands
@@ -55,7 +57,7 @@ func TestCreateChains(t *testing.T) {
 
 	// CreateChains doesn't care for output and we don't any errors
 	// we only want to analize which command generated by the function
-	E := &FakeExecutor{nil, nil, nil}
+	E := &utilexec.FakeExecutor{nil, nil, nil}
 	agent.Helper.Executor = E
 	ip := net.ParseIP("127.0.0.1")
 	fw, _ := NewFirewall(NetIf{"eth0", "A", ip}, &agent)
@@ -81,7 +83,7 @@ func TestDivertTraffic(t *testing.T) {
 	// But because ensureIptablesRule will use same object as
 	// response from os.exec we will see error in test logs,
 	// it's ok as long as function generates expected set of commands.
-	E := &FakeExecutor{nil, errors.New("Rule not found"), nil}
+	E := &utilexec.FakeExecutor{nil, errors.New("Rule not found"), nil}
 	agent.Helper.Executor = E
 	ip := net.ParseIP("127.0.0.1")
 	fw, _ := NewFirewall(NetIf{"eth0", "A", ip}, &agent)
@@ -101,7 +103,7 @@ func TestDivertTraffic(t *testing.T) {
 // firewall rules.
 func TestCreateDefaultRules(t *testing.T) {
 	agent := mockAgent()
-	e := &FakeExecutor{nil, nil, nil}
+	e := &utilexec.FakeExecutor{nil, nil, nil}
 	agent.Helper.Executor = e
 	ip := net.ParseIP("127.0.0.1")
 	fw, _ := NewFirewall(NetIf{"eth0", "A", ip}, &agent)
@@ -118,7 +120,7 @@ func TestCreateDefaultRules(t *testing.T) {
 	}
 
 	agent = mockAgent()
-	e = &FakeExecutor{nil, nil, nil}
+	e = &utilexec.FakeExecutor{nil, nil, nil}
 	agent.Helper.Executor = e
 	fw, _ = NewFirewall(NetIf{"eth0", "A", ip}, &agent)
 
@@ -141,7 +143,7 @@ func TestCreateRules(t *testing.T) {
 	agent := mockAgent()
 
 	// we only care for recorded commands, no need for fake output or errors
-	E := &FakeExecutor{nil, nil, nil}
+	E := &utilexec.FakeExecutor{nil, nil, nil}
 	agent.Helper.Executor = E
 	ip := net.ParseIP("127.0.0.1")
 	fw, _ := NewFirewall(NetIf{"eth0", "A", ip}, &agent)
@@ -166,7 +168,7 @@ func TestCreateU32Rules(t *testing.T) {
 	agent := mockAgent()
 
 	// we only care for recorded commands, no need for fake output or errors
-	E := &FakeExecutor{nil, nil, nil}
+	E := &utilexec.FakeExecutor{nil, nil, nil}
 
 	agent.Helper.Executor = E
 	ip := net.ParseIP("127.0.0.1")
