+++ agent/helpers_test.go
@@ -21,6 +21,9 @@ import (
 	"net"
 	"strings"
 	"testing"
+
+	utilexec "github.com/romana/core/pkg/util/exec"
+	utilos "github.com/romana/core/pkg/util/os"
 )
 
 // TestAppendLineToFile writes data to fake file and ensures it's got written.
@@ -34,13 +37,13 @@ func TestAppendLineToFile(t *testing.T) {
 	   * new(AgentHelperT) returns uninitialized AgentHelperT
 	   * AgentHelperT.Executor not used in this test so we initialize it with
 	     default implementation
-	   * AgentHelperT.OS initialized with FakeOS that implements
+	   * AgentHelperT.OS initialized with utilos.FakeOS that implements
 	     appendFile (used by appendLineToFile)
 	   * Data written by appendLineToFile available via OS.fakeFile.content
 	*/
 
 	// Init fake helpe, &leasefiler
-	fOS := &FakeOS{"", nil}
+	fOS := &utilos.FakeOS{"", nil}
 	helper := Helper{OS: fOS}
 	agent := Agent{Helper: &helper}
 
@@ -50,7 +53,7 @@ func TestAppendLineToFile(t *testing.T) {
 	_ = agent.Helper.appendLineToFile("stub", lease)
 
 	// expect
-	get := fOS.fakeFile.content
+	get := fOS.FakeFile.Content
 	if get != fmt.Sprintf("%s\n", lease) {
 		t.Errorf("AppendLineToFile failed, expect %s\\n, got %q", lease, get)
 	}
@@ -59,8 +62,8 @@ func TestAppendLineToFile(t *testing.T) {
 // TestIsLineInFile is checking that isLineInFile correctly detect presence of
 // given line in the file.
 func TestIsLineInFile(t *testing.T) {
-	// Returning correct lease via FakeOS
-	fOS := &FakeOS{"A 127.0.0.1", nil}
+	// Returning correct lease via utilos.FakeOS
+	fOS := &utilos.FakeOS{"A 127.0.0.1", nil}
 	agent := Agent{Helper: &Helper{OS: fOS}}
 
 	// NetIf to make a lease from
@@ -85,8 +88,8 @@ func TestIsLineInFile(t *testing.T) {
 	// NetIf to make a lease from
 	netif = NetIf{"eth0", "A", ip}
 
-	// Returning wrong lease via FakeOS
-	fOS = &FakeOS{"C 127.0.0.1", nil}
+	// Returning wrong lease via utilos.FakeOS
+	fOS = &utilos.FakeOS{"C 127.0.0.1", nil}
 	agent.Helper.OS = fOS
 	lease = fmt.Sprintf("%s %s", netif.Mac, netif.IP)
 
@@ -110,7 +113,7 @@ func TestDhcpPid(t *testing.T) {
 	// when
 
 	// DhcpPid is expecting pid in stdout
-	E := &FakeExecutor{[]byte("12345"), nil, nil}
+	E := &utilexec.FakeExecutor{[]byte("12345"), nil, nil}
 	agent.Helper.Executor = E
 	out, err := agent.Helper.DhcpPid()
 	// expect
@@ -124,7 +127,7 @@ func TestDhcpPid(t *testing.T) {
 	// when
 
 	// Here testing that DhcpPid's sanity check - pid must be < 65535
-	E = &FakeExecutor{[]byte("1234567"), nil, nil}
+	E = &utilexec.FakeExecutor{[]byte("1234567"), nil, nil}
 	agent.Helper.Executor = E
 	out, err = agent.Helper.DhcpPid()
 	// expect
@@ -140,7 +143,7 @@ func TestIsRouteExist(t *testing.T) {
 	// when
 
 	// IsRouteExist treats non empty output as a success
-	E := &FakeExecutor{[]byte("route exist"), nil, nil}
+	E := &utilexec.FakeExecutor{[]byte("route exist"), nil, nil}
 	agent.Helper.Executor = E
 	ip := net.ParseIP("127.0.0.1")
 	err := agent.Helper.isRouteExist(ip, "32")
@@ -158,7 +161,7 @@ func TestIsRouteExist(t *testing.T) {
 
 	// when
 	// for this test we want to fail isRouteExist by providing nil output
-	E = &FakeExecutor{nil, nil, nil}
+	E = &utilexec.FakeExecutor{nil, nil, nil}
 	agent.Helper.Executor = E
 	err = agent.Helper.isRouteExist(ip, "32")
 
@@ -175,7 +178,7 @@ func TestCreateRoute(t *testing.T) {
 	// when
 
 	// we only care for recorded commands, no need for fake output or errors
-	E := &FakeExecutor{nil, nil, nil}
+	E := &utilexec.FakeExecutor{nil, nil, nil}
 	agent.Helper.Executor = E
 	ip := net.ParseIP("127.0.0.1")
 	_ = agent.Helper.createRoute(ip, "0", "dev", "eth0")
@@ -195,7 +198,7 @@ func TestCreateInterhostRoutes(t *testing.T) {
 	// when
 
 	// we only care for recorded commands, no need for fake output or errors
-	E := &FakeExecutor{nil, nil, nil}
+	E := &utilexec.FakeExecutor{nil, nil, nil}
 	agent.Helper.Executor = E
 	_ = agent.Helper.ensureInterHostRoutes()
